samples/cellular: Replace ftp_client sample with an ftp_put sample
The tele2.net FTP speedtest server seems to no longer be maintained.
Rather than continue to point to a (broken) third-party site,
we have written this new sample, which requires you to have your own
FTP server at the ready.

An implicit assumption with this sample is that, if you are running it
in order to demonstrate the XBee's ability to use FTP,
then you have an FTP server in mind or one that is available.

Instructions to set up such an FTP server are outside the scope
of this sample application.